% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_refs.R
\name{as_refs}
\alias{as_refs}
\title{Generate ref strings digestible by pak}
\usage{
as_refs(dependency_tree)
}
\arguments{
\item{dependency_tree}{data.frame(type, package, version, depth)}
}
\description{
Generate ref strings digestible by pak
}
