% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.base_packages}
\alias{.base_packages}
\title{List of packages that shuold be ignored when fetching dependencies}
\format{
An object of class \code{character} of length 15.
}
\usage{
.base_packages
}
\description{
List of packages that shuold be ignored when fetching dependencies
}
\keyword{datasets}
