% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.description_index}
\alias{.description_index}
\title{Shared index of package descriptions}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.description_index
}
\description{
Shared index of package descriptions
}
\keyword{datasets}
