% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Run logic when package is loaded}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{builtin}

\item{pkgname}{builtin}
}
\description{
Run logic when package is loaded
}
