% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Extract package dependencies from its DESCRIPTION}
\usage{
get_dependencies(description, types = c("Depends", "Imports", "Suggests"))
}
\arguments{
\item{description}{desc::desc}

\item{types}{character()}
}
\description{
Extract package dependencies from its DESCRIPTION
}
