% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{get_description}
\alias{get_description}
\title{Download package tarball and parse the DESCRIPTION}
\usage{
get_description(package, version, cache = .description_index)
}
\arguments{
\item{package}{character(1)}

\item{version}{character(1)}

\item{cache}{mutable env}
}
\description{
Download package tarball and parse the DESCRIPTION
}
