% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solution.R
\name{get_solution}
\alias{get_solution}
\title{Get a complete tree of dependencies for a package.
We are interested in hard+soft dependencies of the package itself,
as well as hard dependencies of all its direct dependencies.}
\usage{
get_solution(
  package,
  version,
  description_cache = .description_index,
  versions_cache = .versions_index,
  recursive_resolution_limit = 1000L
)
}
\arguments{
\item{package}{character(1)}

\item{version}{character(1)}

\item{description_cache}{mutable env}

\item{versions_cache}{mutable env}

\item{recursive_resolution_limit}{integer(1)}
}
\description{
Get a complete tree of dependencies for a package.
We are interested in hard+soft dependencies of the package itself,
as well as hard dependencies of all its direct dependencies.
}
