% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{Get a complete tree of dependencies for a package.
We are interested in hard+soft dependencies of the package itself,
as well as hard dependencies of all its direct dependencies.}
\usage{
get_tree(
  package,
  version,
  description_cache = .description_index,
  versions_cache = .versions_index
)
}
\arguments{
\item{package}{character(1)}

\item{version}{character(1)}

\item{description_cache}{mutable env}

\item{versions_cache}{mutable env}
}
\description{
Get a complete tree of dependencies for a package.
We are interested in hard+soft dependencies of the package itself,
as well as hard dependencies of all its direct dependencies.
}
