% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_refs.R
\name{make_refs}
\alias{make_refs}
\title{Generate ref strings digestible by pak}
\usage{
make_refs(dependency_tree)
}
\arguments{
\item{dependency_tree}{data.frame(type, package, version, depth)}
}
\description{
Generate ref strings digestible by pak
}
