% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.versions_index}
\alias{.versions_index}
\title{Shared index of package versions}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.versions_index
}
\description{
Shared index of package versions
}
\keyword{datasets}
