% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_versions.R
\name{get_versions}
\alias{get_versions}
\title{Get a list of package versions
TODO: work beyond r-project's mirror of CRAN
TODO: work beyond CRAN}
\usage{
get_versions(package, cache = .versions_index)
}
\arguments{
\item{package}{character(1)}

\item{cache}{mutable env}
}
\description{
Get a list of package versions
TODO: work beyond r-project's mirror of CRAN
TODO: work beyond CRAN
}
