% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_versions.R
\name{match_versions}
\alias{match_versions}
\title{For each package dependency assign the latest version
that was avaiable before package release}
\usage{
match_versions(description, deps, index = .versions_index)
}
\arguments{
\item{description}{desc::desc}

\item{deps}{data.frame(type, package, version)}

\item{index}{mutable env}
}
\description{
For each package dependency assign the latest version
that was avaiable before package release
}
